Branch displacement is calculated as 32 bits.

+++ src/input/x86/translators/branch.cpp
@@ -20,7 +20,7 @@ void branch_translator::do_translate()
 		auto next_target_node = builder().insert_add(builder().insert_read_pc()->val(), builder().insert_constant_u64(instruction_length)->val());
 		builder().insert_write_mem(new_rsp->val(), next_target_node->val());
 
-		int32_t value = xed_decoded_inst_get_branch_displacement(xed_inst());
+		int64_t value = xed_decoded_inst_get_branch_displacement(xed_inst());
 		uint64_t target = value + instruction_length;
 
 		auto target_node = builder().insert_add(builder().insert_read_pc()->val(), builder().insert_constant_u64(target)->val());
@@ -46,7 +46,7 @@ void branch_translator::do_translate()
 
 	case XED_ICLASS_JMP: {
 		xed_uint_t instruction_length = xed_decoded_inst_get_length(xed_inst());
-		int32_t branch_displacement = xed_decoded_inst_get_branch_displacement(xed_inst());
+		int64_t branch_displacement = xed_decoded_inst_get_branch_displacement(xed_inst());
 		uint64_t branch_target = branch_displacement + instruction_length;
 
 		auto target = builder().insert_add(builder().insert_read_pc()->val(), builder().insert_constant_u64(branch_target)->val());
